README: Fix license example
LGPL2-only (and LGPL2.1-only) works with GPL3, because of clause 3 of
the LGPL, which says "If a newer version than version 2 of the ordinary
GNU General Public License has appeared, then you can specify that
version instead if you wish.".

The case that doesn't work: GPL2-only with LGPL3 or later, because the
LGPL3 only allows conversion to the GPL3, not the GPL2.